Unify fly machines <command> <mid> commands
With the exception of clone, which still requires an app to be
specified, machines subcommands that take machine IDs will now use the
app of the first machine specified. All subsequent machines are required
to be in the same app. Notably, this

* loosens the requirements for `leases` and
* strengthens the requirements for `start` and `stop`.

As part of this, all app/machine selection and setup logic (including
creation of the flaps client for the app selected) has been consolidated
to select.go. This will make it easier to implement a '--select' flag
for machine commands. In the meantime, though, it also

* eliminates calls to apps.BuildContext that establish an agent even
  when unnecessary and
* creates flaps clients based on app name instead of an AppCompact,
  eliminating an API round-trip in many cases.