Add --select for machine commands
When this flag is used, users can choose the machine(s) on which to
operate from a menu. This improves accessibility and the UX when working
with machines.

When --select is used, no machine IDs may be specified as arguments, and
an app name must be provided (through fly.toml, FLY_APP, or -a).